Разделение на модули, подключение точки входа к index.html
