+++ js/constants.js
@@ -1,0 +1,19 @@
+// Типы сдаваемой жилплощади
+const TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+
+//Время въезда/выезда
+const CHECK_IN_OUT = ['12:00', '13:00', '14:00'];
+
+//Удобства
+const FEATURES = ['wi-fi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+
+//Фото
+const PHOTOS = ['https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg', 'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg', 'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'];
+
+//Массив с заголовками объявлений
+const OFFER_TITLES = ['Предлагаю для сдачи', 'Аренда', 'Лучшая цена'];
+
+//Массив с описанием объявления
+const OFFER_DESCRIPTIONS = ['Рядом с центром', 'Необходимая инфраструктура', 'Тихий район', 'Рядом с парком'];
+
+export {TYPES, CHECK_IN_OUT, FEATURES, PHOTOS, OFFER_TITLES, OFFER_DESCRIPTIONS};
